requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
PyPDF2>=3.0.0
